% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nudibranch.R
\name{geom_nudibranch}
\alias{geom_nudibranch}
\title{geom_nudibranch}
\usage{
geom_nudibranch(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  nudge_x = 0,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{inherit.aes}{logical, whether inherit aes from ggplot()}

\item{na.rm}{logical, whether remove NA values}

\item{by}{one of 'width' or 'height'}

\item{nudge_x}{horizontal adjustment to nudge nudibranchs}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer adding nudibranchs
}
\examples{
library("ggplot2")
ggplot(mtcars) +
geom_nudibranch(aes(mpg, wt), nudibranch = "O.rosacea", size = 5)

set.seed(1)
df <- data.frame(x = rnorm(10),
                 y = rnorm(10),
                 image = sample(c("O.rosacea",
                                  "H.crassicornis",
                                  "S.shawl",
                                  "Sea.lemon",
                                  "T.catalinae"),
                                 size = 10, replace = TRUE))
 ggplot(df) +
geom_nudibranch(aes(x, y, nudibranch = image), size = 5)

}
